ord-lsp: fold helper files into corresponding parent files
